refactor: 주간 분석 생성을 위해 일일 통계치를 얻는 로직을 DailyService 가 아닌 DailyRepo 에서 찾도록 변경
V1 버전
