+++ src/main/java/site/radio/report/weekly/service/WeeklyFacade.java
@@ -5,6 +5,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -13,8 +14,10 @@ import site.radio.clova.dailyReport.DailyReportExtractor;
 import site.radio.clova.dailyReport.DailyReportPreprocessor;
 import site.radio.clova.dto.ClovaResponseDto;
 import site.radio.clova.service.ClovaService;
+import site.radio.clova.weekly.ClovaWeeklyReportRequestDto;
 import site.radio.letter.Letter;
 import site.radio.letter.LetterService;
+import site.radio.report.daily.domain.DailyReport;
 import site.radio.report.daily.service.DailyReportService;
 import site.radio.report.weekly.dto.WeeklyReportResponseDto;
 
@@ -50,6 +53,27 @@ public class WeeklyFacade {
         // 클로바로부터 받은 일일 리포트의 분석결과들을 저장
         dailyReportService.saveClovaDailyAnalysisResult(lettersByAnalysisResult);
 
-        return null;
+        // 주간 분석에 필요한 위로 한마디 구하기
+        List<LocalDate> oneWeek = createOneWeek(startDate);
+        List<DailyReport> dailyReports = dailyReportService.getDailyReportsOfOneWeek(userId, oneWeek);
+        ClovaWeeklyReportRequestDto clovaWeeklyReportRequestDto = ClovaWeeklyReportRequestDto.from(combineDescription(dailyReports));
+
+        // 주간 분석 요청 및 응답
+        ClovaResponseDto clovaResponseDto = clovaService.sendWeeklyReportRequest(clovaWeeklyReportRequestDto);
+        String resultMessage = clovaResponseDto.getResultMessage();
+
+        return weeklyReportService.createWeeklyReport(userId, oneWeek, resultMessage);
+    }
+
+    private List<LocalDate> createOneWeek(LocalDate startDate) {
+        return IntStream.range(0, 7)
+                .mapToObj(startDate::plusDays)
+                .toList();
+    }
+
+    private String combineDescription(List<DailyReport> dailyReports) {
+        return dailyReports.stream()
+                .map(DailyReport::getDescription)
+                .collect(Collectors.joining());
     }
 }
